frontend/api/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
pdfminer.six
transformers
torch
safetensors
pydantic
numpy
scikit-learn 